requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
google-generativeai>=0.8.0,<1.0.0
google-api-core>=2.0.0,<3.0.0
openai>=1.50.0,<2.0.0
httpx>=0.27.0,<1.0.0
Pillow>=10.0.0,<11.0.0
structlog>=24.0.0,<25.0.0
python-multipart>=0.0.9,<1.0.0
